Added authentication

+++ apps/server/src/modules/mood/controller.ts
@@ -1,4 +1,4 @@
-import { controller } from "../../controllers/controller";
+import { controller } from "../../common/controller";
 import { getValidModel } from "../../common/validation";
 import { AddMoodRequestSchema, DeleteMoodRequestSchema } from "./schema";
 import { moodService } from "./service";
@@ -7,7 +7,7 @@ export const moodController = {
   addMoodEntry: controller(async (req) => {
     const moodParams = getValidModel(AddMoodRequestSchema, req.body);
 
-    const addedMood = await moodService.addMoodEntry({ ...moodParams, userId: "1" });
+    const addedMood = await moodService.addMoodEntry({ ...moodParams, userId: req.user.id });
 
     return { status: 200, result: { mood: addedMood } };
   }),
@@ -15,15 +15,13 @@ export const moodController = {
   deleteMoodEntry: controller(async (req) => {
     const moodParams = getValidModel(DeleteMoodRequestSchema, req.body);
 
-    await moodService.deleteMoodEntry(moodParams.id);
+    await moodService.deleteMoodEntry({ entryId: moodParams.id, userId: req.user.id });
 
     return { status: 200, result: { message: "Successfully deleted the mood" } };
   }),
 
   listMoodEntries: controller(async (req) => {
-    // const moodParams = getValidModel(DeleteMoodRequestSchema, req.query);
-
-    const entries = await moodService.listMoodEntries({ userId: "1" });
+    const entries = await moodService.listMoodEntries({ userId: req.user.id });
 
     return { status: 200, result: { moods: entries } };
   }),
